파티클 GS 적용해서 그리기

+++ DX11_Wizard_of_Legend/ENGINE_SOURCE/jsEntity.cpp
@@ -3,11 +3,11 @@
 namespace js
 {
 	Entity::Entity()
-		: mID((UINT32)reinterpret_cast<UINT32>(this))
+		: mID(reinterpret_cast<UINT32>(this))
 	{
 	}
 	Entity::Entity(const Entity& other)
-		: mID((UINT32)reinterpret_cast<UINT32>(this))
+		: mID(reinterpret_cast<UINT32>(this))
 		, mName(other.mName)
 	{
 	}
